requirements.txt
apache-beam==2.61.0
apache-flink==2.2.0
apache-flink-libraries==2.2.0
attrs==25.4.0
avro==1.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
cloudpickle==2.2.1
confluent-kafka==2.13.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.1
fasteners==0.20
find_libpython==0.5.1
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.11
jsonpickle==3.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
numpy==2.1.3
objsize==0.7.1
orjson==3.11.7
packaging==26.0
pandas==2.2.3
pemja==0.5.5
proto-plus==1.27.1
protobuf==5.29.6
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydot==1.4.2
pyflink==1.0
PyJWT==2.11.0
pymongo==4.16.0
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
ruamel.yaml==0.19.1
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zstandard==0.25.0